1. Модульные тесты. [#414589]

+++ test/ru/job4j/ConverterTest.java
@@ -13,11 +13,12 @@ public class ConverterTest {
         Assert.assertEquals(expected, out);
     }
 
+
     @Test
     public void whenConvert180RblThen3Dlr() {
         int in = 180;
         int expected = 3;
-        int out = Converter.rubleToDollar(in);
+        float out = Converter.rubleToDollar(in);
         Assert.assertEquals(expected, out);
     }
 }
